feat: clean secrets in logs
In the previous version, all messages were logged to the console via the
default JavaScript `console.log`. With this logging method, if some
error occurs during the interaction with a webhook, its content is
logged in the console as part of the error object.

In this version, we added Winston logger to the project with the logic
of cleaning secrets in logs. All webhook tokens are now added to the
secrets after the reading of the config file is completed.

Extend info that is printed to the log in case of errors. For
network-related errors print response status and request data, for all
errors print their stack trace.